Added the ablity to pick classes in assignment
