+++ app/src/main/res/layout/fragment_dropdown_item.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TextView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/textView"
+    android:layout_width="match_parent"
+    android:textColor="@color/white"
+    android:textStyle="bold"
+    android:layout_height="wrap_content"
+    android:text="TextView" />
