+++ app/src/main/res/layout/fragment_enter_assignment.xml
@@ -17,11 +17,11 @@
         android:layout_marginBottom="16dp"
         android:clickable="true"
         android:onClick="buildAssignment"
-        android:contentDescription="Finish"
+        android:contentDescription="@string/finish"
         android:src="@drawable/ic_baseline_check_24"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        />
+        android:focusable="true" />
 
     <EditText
         android:id="@+id/assignmentName"
@@ -38,6 +38,22 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/dropdown"
+        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.ExposedDropdownMenu"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintStart_toStartOf="@+id/assignmentName"
+        app:layout_constraintTop_toBottomOf="@+id/assignmentName">
+
+        <AutoCompleteTextView
+            android:id="@+id/class_choice"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Class"
+            android:textColor="@color/white" />
+    </com.google.android.material.textfield.TextInputLayout>
+
     <com.pagewisegroup.pagewise.DateDisplayView
         android:id="@+id/assignmentDueDate"
         android:layout_width="wrap_content"
@@ -51,7 +67,7 @@
         android:text="Due: MM/DD/YYYY"
         android:textColor="@android:color/white"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/assignmentName" />
+        app:layout_constraintTop_toBottomOf="@+id/dropdown" />
 
     <EditText
         android:id="@+id/pageStart"
