+++ app/src/main/res/values/strings.xml
@@ -16,6 +16,7 @@
 
     <string name="hello_first_fragment">Hello first fragment</string>
     <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
+    <string name="finish">Finish</string>
 
     <!--for student view spinner-->
     <string-array name="StudentActions">
